mupl: create song API

+++ src/main/java/com/mupl/music_service/service/impl/SongServiceImpl.java
@@ -1,6 +1,7 @@
 package com.mupl.music_service.service.impl;
 
-import com.mupl.music_service.dto.request.SongCreateRequest;
+import com.mupl.music_service.dto.request.SongRequest;
+import com.mupl.music_service.dto.response.PageableResponse;
 import com.mupl.music_service.dto.response.SongResponse;
 import com.mupl.music_service.entity.SongEntity;
 import com.mupl.music_service.exception.BadRequestException;
@@ -13,8 +14,11 @@ import com.mupl.music_service.utils.constain.EventType;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Schedulers;
@@ -34,20 +38,20 @@ public class SongServiceImpl implements SongService {
     private final StorageService storageService;
 
     @Override
-    public Mono<SongResponse> createSong(SongCreateRequest songCreateRequest) {
-        SongEntity songEntity = modelMapper.map(songCreateRequest, SongEntity.class);
+    public Mono<SongResponse> createSong(SongRequest songRequest) {
+        SongEntity songEntity = modelMapper.map(songRequest, SongEntity.class);
         songEntity.setCreatedAt(LocalDateTime.now());
         songEntity.setUpdatedAt(LocalDateTime.now());
         return songRepository.save(songEntity)
                 .flatMap(entity -> Mono.when(
-                                createArtistSongRelationship(entity.getSongId(), songCreateRequest.getArtistIds()),
-                                createGenreSongRelationship(entity.getSongId(), songCreateRequest.getGenreIds()))
+                                createArtistSongRelationship(entity.getSongId(), songRequest.getArtistIds()),
+                                createGenreSongRelationship(entity.getSongId(), songRequest.getGenreIds()))
                         .thenReturn(modelMapper.map(songEntity, SongResponse.class))
                         .doOnSuccess(response -> {
-                            storageService.uploadFile(response.getSongId(), songCreateRequest.getImageFile(), EventType.IMAGE_UPLOADED)
+                            storageService.uploadFile(response.getSongId(), songRequest.getImageFile(), EventType.IMAGE_UPLOADED)
                                     .subscribeOn(Schedulers.boundedElastic())
                                     .subscribe();
-                            storageService.uploadFile(response.getSongId(), songCreateRequest.getSongFile(), EventType.SONG_UPLOADED)
+                            storageService.uploadFile(response.getSongId(), songRequest.getSongFile(), EventType.SONG_UPLOADED)
                                     .subscribeOn(Schedulers.boundedElastic())
                                     .subscribe();
                         })
@@ -76,7 +80,66 @@ public class SongServiceImpl implements SongService {
                 });
     }
 
+    @Override
+    public Mono<PageableResponse> getSongs(Pageable pageable, String albumId) {
+        Flux<SongEntity> flux;
+        if (StringUtils.isNotBlank(albumId)) {
+            flux = songRepository.findAllByAlbumId(Integer.parseInt(albumId), pageable);
+        } else {
+            flux = songRepository.findAllBy(pageable);
+        }
+        return flux
+                .map(song -> modelMapper.map(song, SongResponse.class))
+                .collectList()
+                .zipWith(songRepository.count())
+                .map(tuple -> new PageableResponse(tuple.getT1(), pageable, tuple.getT2()));
+    }
+
+    @Override
+    public Mono<PageableResponse> getSongs(Pageable pageable, List<String> ids) {
+        List<Integer> mapIds = ids.stream().map(Integer::parseInt).toList();
+        return songRepository.findByGenres(mapIds, pageable)
+                .map(song -> modelMapper.map(song, SongResponse.class))
+                .collectList()
+                .zipWith(songRepository.count())
+                .map(tuple -> new PageableResponse(tuple.getT1(), pageable, tuple.getT2()));
+
+    }
+
+    @Override
+    public Mono<SongResponse> updateSong(String songId, SongRequest songRequest) {
+        long id = Long.parseLong(songId);
+        return songRepository.findById(id)
+                .flatMap(entity -> {
+                    entity.setReleasedAt(songRequest.getReleasedAt());
+                    entity.setTitle(songRequest.getTitle());
+                    entity.setAlbumId(songRequest.getAlbumId());
+                    entity.setIsFreeToPlay(songRequest.getIsFreeToPlay());
+                    return songRepository.save(entity)
+                            .map(song -> modelMapper.map(song, SongResponse.class));
+                })
+                .doOnSuccess(response -> {
+                    storageService.deleteObject(response.getImagePath())
+                            .then(Mono.defer(() -> storageService.uploadFile(id, songRequest.getImageFile(), EventType.IMAGE_UPLOADED)))
+                            .subscribeOn(Schedulers.boundedElastic())
+                            .subscribe();
+                    storageService.deleteObject(response.getSongPath())
+                            .then(Mono.defer(() -> storageService.uploadFile(id, songRequest.getSongFile(), EventType.SONG_UPLOADED)))
+                            .subscribeOn(Schedulers.boundedElastic())
+                            .subscribe();
+                    createArtistSongRelationship(id, songRequest.getArtistIds())
+                            .subscribeOn(Schedulers.boundedElastic())
+                            .subscribe();
+                    createGenreSongRelationship(id, songRequest.getGenreIds())
+                            .subscribeOn(Schedulers.boundedElastic())
+                            .subscribe();
+                });
+    }
+
     private Mono<Void> createArtistSongRelationship(Long songId, List<Integer> artistIds) {
+        if (CollectionUtils.isEmpty(artistIds)) {
+            return Mono.empty();
+        }
         return artistSongService.deleteAllBySongId(songId)
                 .then(Mono.defer(() -> Flux.fromIterable(artistIds)
                         .flatMap(artistId -> artistSongService.create(artistId, songId))
@@ -84,6 +147,9 @@ public class SongServiceImpl implements SongService {
     }
 
     private Mono<Void> createGenreSongRelationship(Long songId, List<Integer> genreIds) {
+        if (CollectionUtils.isEmpty(genreIds)) {
+            return Mono.empty();
+        }
         return genreSongService.deleteAllBySongId(songId)
                 .then(Mono.defer(() -> Flux.fromIterable(genreIds)
                         .flatMap(genreId -> genreSongService.create(genreId, songId))
